Clean code, better management, fix small bugs

+++ Game_running_package/main.py
@@ -1,7 +1,9 @@
 import pygame
-from enemy import Enemy
-from player import Player
-from sprites import enemies, bullets, players, bonuses
+
+from Enemies_package.asteroid import Asteroid
+from Enemies_package.enemy import Enemy
+from Sprites_package.sprites import players, asteroids, guns, bonuses
+from Player_package.player import Player
 
 pygame.init()
 
@@ -13,13 +15,13 @@ screen = pygame.display.set_mode([SCREEN_WIDTH, SCREEN_HEIGHT])
 pygame.mouse.set_visible(False)
 
 # Menu Audio
-background_audio = pygame.mixer.Sound("Audio/Background_menu_v2.mp3")
+background_audio = pygame.mixer.Sound("Additional_resources/Audio/Background_menu_v2.mp3")
 background_audio.play(loops=-1)
 background_audio.set_volume(0.2)
 background_audio.play()
 
 # Game messages
-font = pygame.font.Font('Font/Pixel_font.ttf', 50)
+font = pygame.font.Font('Additional_resources/Font/Pixel_font.ttf', 50)
 
 play_again_m = font.render('Press SPACE to play again!', False, (0, 0, 0))
 play_again_mR = play_again_m.get_rect(center=(400, 200))
@@ -35,8 +37,8 @@ player = Player()
 players.add(player)
 
 for i in range(4):
-    enemy = Enemy()
-    enemies.add(enemy)
+    asteroid = Asteroid()
+    asteroids.add(asteroid)
 
 clock = pygame.time.Clock()
 game_over = False
@@ -45,9 +47,9 @@ first_run = True
 
 
 # Display score
-def display_score(score):
-    score_surf = font.render(f'Score: {score}', False, (64, 64, 64))
-    score_rect = score_surf.get_rect(center=(400, 50))
+def display_score(score, pos_x, pos_y, color):
+    score_surf = font.render(f'Score: {score}', False, color)
+    score_rect = score_surf.get_rect(center=(pos_x, pos_y))
     screen.blit(score_surf, score_rect)
 
 
@@ -58,40 +60,38 @@ while not game_over:
         if event.type == pygame.KEYDOWN and event.key == pygame.K_ESCAPE:
             game_over = True
 
-        if gameplay_state:
-            print(".")
-        else:
-            if event.type == pygame.KEYDOWN and event.key == pygame.K_SPACE:
-                bullets.empty()
-                enemies.empty()
-                bonuses.empty()
-                for i in range(4):
-                    enemy = Enemy()
-                    enemies.add(enemy)
-                    player.kill()
-                    player = Player()
-                    players.add(player)
+        if event.type == pygame.KEYDOWN and event.key == pygame.K_SPACE and not gameplay_state:
+            guns.empty()
+            asteroids.empty()
+            bonuses.empty()
 
-                gameplay_state = True
+            for i in range(4):
+                asteroid = Asteroid()
+                asteroids.add(asteroid)
+
+            player.kill()
+            player = Player()
+            players.add(player)
+            gameplay_state = True
 
     if gameplay_state:
         screen.fill('#0E034E')
-        display_score(player.score)
+        display_score(player.score, 400, 50, '#E7CFCF')
 
         # Draw the sprites
         players.draw(screen)
-        enemies.draw(screen)
-        bullets.draw(screen)
+        asteroids.draw(screen)
+        guns.draw(screen)
         bonuses.draw(screen)
 
         # Update the sprites
         players.update()
-        enemies.update()
-        bullets.update()
+        asteroids.update()
+        guns.update()
         bonuses.update()
 
         # Check for collisions between the player and enemies
-        collided_player = pygame.sprite.spritecollide(player, enemies, False)
+        collided_player = pygame.sprite.spritecollide(player, asteroids, False)
         if collided_player:
             first_run = False
             gameplay_state = False
@@ -105,6 +105,7 @@ while not game_over:
         else:
             screen.fill('#3C00AD')
             screen.blit(play_again_m, play_again_mR)
+            display_score(player.score, 400, 700, '#000000')
 
     # Update the screen
     pygame.display.flip()
